+++ src/command_modules/azure-cli-sql/HISTORY.rst
@@ -7,6 +7,7 @@ Unreleased
 +++++++++++++++++++
 * Adding support for SQL Transparent Data Encryption (TDE) and TDE with Bring Your Own Key
 * Added az sql db list-deleted command and az sql db restore --deleted-time parameter, allowing the ability to find and restore deleted databases.
+* Added az sql db op list and az sql db op cancel, allowing the ability to list and cancel in-progress operations on database.
 
 2.0.12 (2017-09-22)
 +++++++++++++++++++