+++ src/command_modules/azure-cli-sql/HISTORY.rst
@@ -3,7 +3,7 @@
 Release History
 ===============
 
-Unreleased
+2.0.13 (2017-10-09)
 +++++++++++++++++++
 * Adding support for SQL Transparent Data Encryption (TDE) and TDE with Bring Your Own Key
 * Added az sql db list-deleted command and az sql db restore --deleted-time parameter, allowing the ability to find and restore deleted databases.